Refactoring to an abstract Buffer class and concrete CPU and GPU implementations of those
